Renamed clinical trials file and corrected some spelling errors in NSF file
