+++ trials.py
